Add the git-retimestamp script
The script resets the timestamps of all commits in the currently checked
out branch to the current time.

This is done to ensure that GitHub displays the commits in pull requests
in the correct order (i.e. according to the parent-child relationship of
the commits and not according to the time when a commit was originally
made).